fix bugs, add new attributes and refactor codes.
