+++ README.md
@@ -9,6 +9,8 @@ TwinklingRefreshLayout延伸了Google的SwipeRefreshLayout的思想,不在列表
 5. setOnRefreshListener中拥有大量可以回调的方法
 6. 将Header和Footer抽象成了接口,并回调了滑动过程中的系数,方便实现个性化的Header和Footer
 
+![](art/pic_large.png)
+
 ## Demo
 [下载Demo](art/app-debug.apk)
