+++ app/src/main/res/layout/fragment_gridview.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/refresh"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -11,4 +11,4 @@
         android:background="@android:color/white"
         android:columnWidth="200dp"
         android:numColumns="2" />
-</com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout>
+</com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout>
