+++ app/src/main/res/layout/fragment_listview.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/refresh"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -10,4 +10,4 @@
         android:layout_height="match_parent"
         android:background="@android:color/white"
         android:dividerHeight="0dp"/>
-</com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout>
+</com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout>
