+++ app/src/main/res/layout/fragment_recycler.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/refresh"
     android:layout_width="match_parent"
@@ -12,4 +12,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#fff"/>
-</com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout>
+</com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout>
