+++ app/src/main/res/layout/fragment_webview.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/refreshLayout"
     android:background="#fff"
     android:layout_width="match_parent"
@@ -10,4 +10,4 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="#fff" />
-</com.lcodecore.tkrefreshlayout.v3.TwinklingRefreshLayout>
+</com.lcodecore.tkrefreshlayout.TwinklingRefreshLayout>
